Fix in CreateInstance... + some work on simple addition of hierarchies for a list of idables
